% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_contributors.R
\name{github_contributors}
\alias{github_contributors}
\title{GitHub contributors}
\usage{
github_contributors(
  repos,
  agg_by_repo = TRUE,
  agg_by_user = TRUE,
  token = gh::gh_token(),
  sort_rows = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{token}{GitHub authentication token. 
See \link[gh]{gh} documentation for more details.}

\item{verbose}{Print messages.}
}
\description{
Get contributors to each GitHub repository.
}
\examples{
repos <- github_org(org = "neurogenomics", field = "repos")
contributors <- github_contributors(repos = repos)
}
