% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_metadata.R
\name{github_metadata}
\alias{github_metadata}
\title{GitHub metadata}
\usage{
github_metadata(repos, field, token = gh::gh_token(), verbose = TRUE)
}
\arguments{
\item{field}{Field to search.}

\item{token}{GitHub authentication token. 
See \link[gh]{gh} documentation for more details.}

\item{verbose}{Print messages.}
}
\description{
Extract further GitHub metadata from the output out
 an initial search with \link[SkillNet]{github_org}.
}
\keyword{internal}
