% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_network.R
\name{github_network}
\alias{github_network}
\title{GitHub network}
\source{
\href{https://docs.github.com/en/rest/repos#get}{GitHub API: Repositories}
\href{https://docs.github.com/en/rest/projects/cards}{GitHub API: Projects (classic)}
\href{https://docs.github.com/en/enterprise-cloud\\@latest/issues/planning-and-tracking-with-projects/automating-your-project/using-the-api-to-manage-projects}{GitHub API: ProjectsV2}
\href{https://github.com/ChadGoymer/githapi}{
Non-functional GH package with similar objective}
}
\usage{
github_network(
  org,
  report = NULL,
  add_contributors = TRUE,
  add_issues = FALSE,
  add_comments = FALSE,
  token = gh::gh_token(),
  verbose = TRUE
)
}
\arguments{
\item{org}{GitHub organization name.}

\item{add_contributors}{Extract metadata about which users 
have contributed to which repos.}

\item{add_issues}{Extract metadata from Issues within each GitHub repo.}

\item{add_comments}{Extract metadata from Comments within each GitHub Issue.}

\item{token}{GitHub authentication token. 
See \link[gh]{gh} documentation for more details.}

\item{verbose}{Print messages.}
}
\value{
igraph object
}
\description{
Construct a network from a GitHub Organization's 
repositories and contributors.
}
\examples{
g <- github_network(org = "neurogenomics")
}
