% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_network_plot.R
\name{github_network_plot}
\alias{github_network_plot}
\title{GitHub network plot}
\usage{
github_network_plot(
  graph,
  shape = c("image", "hexagon"),
  layout = "layout_nicely",
  physics = FALSE,
  show_plot = list(r = TRUE, browser = TRUE),
  save_path = NULL,
  width = NULL,
  height = NULL,
  background = "white",
  verbose = TRUE
)
}
\arguments{
\item{graph}{An \link[igraph]{graph} generated by 
\link[SkillNet]{github_network}.}

\item{shape}{: String. Default to 'ellipse'. The shape defines what the node looks like. There are two types of nodes. One type has the label inside of it and the other type has the label underneath it. The types with the label inside of it are: ellipse, circle, database, box, text. The ones with the label outside of it are: image, circularImage, diamond, dot, star, triangle, triangleDown, hexagon, square and icon.}

\item{layout}{: Character Name of igraph layout function to use. Default to "layout_nicely"}

\item{physics}{: Boolean. Default to true. When false, the node is not part of the physics simulation. It will not move except for from manual dragging.}

\item{show_plot}{A named list with two items:
\itemize{
\item{r}{Whether to show the dependency graph in R/Rstudio
(will \emph{not} show user-specific \code{background} color)}.
\item{browser}{Whether to show the dependency graph in default web browser
(will show user-specific \code{background} color)}.
}}

\item{save_path}{Path to save the plot to, as an interactive,
self-container HTML file.}

\item{width}{: String. Default to "100\%". The width of the network in pixels or as a percentage.}

\item{height}{: String. Default to "100\%". The height of the network in pixels or as a percentage.}

\item{background}{: Text string giving the html background color of the widget. Defaults to white.}

\item{verbose}{Print messages.}
}
\description{
Generate an interactive network plot for all repositories 
in a GitHub Organization, where users and repositories are nodes and
 contributions are edges.
Edge width is proportional to the number of contributions 
 each user made to a given repository.
User nodes use images automatically pulled from 
 their respective GitHub profile.
Repo node size is proportional the size of the repository.
}
\examples{
g <- github_network(org = "neurogenomics")
vis <- github_network_plot(graph = g$graph)
}
