% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_network_vertices.R
\name{github_network_vertices}
\alias{github_network_vertices}
\title{GitHub network: vertices}
\usage{
github_network_vertices(report)
}
\description{
Create GitHub network vertices.
}
\keyword{internal}
