% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_org.R
\name{github_org}
\alias{github_org}
\title{GitHub Organization}
\usage{
github_org(
  org,
  field = NULL,
  token = gh::gh_token(),
  as_datatable = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{field}{Field to search.}

\item{token}{GitHub authentication token. 
See \link[gh]{gh} documentation for more details.}

\item{verbose}{Print messages.}
}
\description{
Get GitHub Organization metadata.
}
