% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_report.R
\name{github_report}
\alias{github_report}
\title{GitHub report}
\usage{
github_report(
  org,
  add_contributors = TRUE,
  add_issues = FALSE,
  add_comments = FALSE,
  token = gh::gh_token(),
  verbose = TRUE
)
}
\arguments{
\item{org}{GitHub organization name.}

\item{add_contributors}{Extract metadata about which users 
have contributed to which repos.}

\item{add_issues}{Extract metadata from Issues within each GitHub repo.}

\item{add_comments}{Extract metadata from Comments within each GitHub Issue.}

\item{token}{GitHub authentication token. 
See \link[gh]{gh} documentation for more details.}

\item{verbose}{Print messages.}
}
\value{
A nested list of \link[data.table]{data.table}s.
}
\description{
Generate a report from all of a GitHub Organization's repositories.
}
\examples{
report <- github_report(org = "neurogenomics")
}
